requirements.txt
Pillow==5.4.1
pywin32==224
